Project Activity Mapping

+++ src/Modules/Execution/Execution.Application/Features/ProjectActivities/Commands/ProjectActivityCommands.cs
@@ -4,5 +4,5 @@ using Himapp.Execution.Application.Features.ProjectActivities.Models;
 namespace Himapp.Execution.Application.Features.ProjectActivities.Commands;
 
 public sealed record CreateProjectActivityCommand(CreateProjectActivityRequest Request) : IRequest<ProjectActivityModel>;
-public sealed record UpdateProjectActivityCommand(long Id, UpdateProjectActivityRequest Request) : IRequest<ProjectActivityModel?>;
-public sealed record DeleteProjectActivityCommand(long Id) : IRequest<bool>;
+public sealed record UpdateProjectActivityCommand(int Id, UpdateProjectActivityRequest Request) : IRequest<ProjectActivityModel?>;
+public sealed record DeleteProjectActivityCommand(int Id) : IRequest<bool>;
